fast-transcription/requirements.txt
requests==2.32.3
python-dotenv==1.1.0
mutagen==1.47.0
